Nbrb: do not work export to file, select rates of the period
